Adding dipext-sock_r3.
Converted the r2 altium pcb files to kicad. Rewired the traces to match the flipping of the connections similar to the corresponding changes in dipext-pin R6. Note: had to suspend the DRC checks, since the schematic did not import and so had to route againt the netlist thats in the pcb file. Note2: the "import non-kicad" option  to import altium files only shows up in pcbnew if you start pcbnew directly, and not through the project explorer of kicad."